python/pyproject.toml
[project]
name = "retrace-sdk"
version = "0.1.7"
description = "Record, replay, fork & share AI agent executions"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Yash Bogam"}]
keywords = ["ai", "agent", "tracing", "observability", "replay", "llm", "openai", "anthropic", "gemini"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "websocket-client>=1.9.0",
    "requests>=2.32.0",
]

[project.urls]
Homepage = "https://retrace.yashbogam.me"
Repository = "https://github.com/yash1511-bogam/retrace-sdk-python"
Documentation = "https://retrace.yashbogam.me/docs/sdk-python"

[project.optional-dependencies]
gemini = ["google-genai>=1.52.0"]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.30.0"]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "mypy>=1.10", "ruff>=0.5"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.hatch.build.targets.wheel]
packages = ["src/retrace"]
